Struct utf8::LossyDecoder

pub struct LossyDecoder<F: FnMut(&str)> {
    // some fields omitted
}

A push-based, lossy decoder for UTF-8. Errors are replaced with the U+FFFD replacement character.

Users “push” bytes into the decoder, which in turn “pushes” &str slices into a callback.

For example, String::from_utf8_lossy (but returning String instead of Cow) can be rewritten as:

fn string_from_utf8_lossy(input: &[u8]) -> String {
    let mut string = String::new();
    utf8::LossyDecoder::new(|s| string.push_str(s)).feed(input);
    string
}

Note: Dropping the decoder signals the end of the input: If the last input chunk ended with an incomplete byte sequence for a code point, this is an error and a replacement character is emitted. Use std::mem::forget to inhibit this behavior.

Methods

impl<F: FnMut(&str)> LossyDecoder<F>

fn new(push_str: F) -> Self

Create a new decoder from a callback.

fn feed(&mut self, input: &[u8])

Feed one chunk of input into the decoder.

The input is decoded lossily and the callback called once or more with &str string slices.

If the UTF-8 byte sequence for one code point was split into this bytes chunk and previous bytes chunks, it will be correctly pieced back together.