Crate utf8

Structs

Decoder
A low-level, zero-copy UTF-8 decoder with error handling.

InlineString
Like String, but does not allocate memory and has a fixed capacity of 4 bytes. This is used by Decoder to represent either the empty string or a single code point.

LossyDecoder
A push-based, lossy decoder for UTF-8. Errors are replaced with the U+FFFD replacement character.

Enums

Result

Constants

REPLACEMENT_CHARACTER
The replacement character, U+FFFD. In lossy decoding, insert it for every decoding error.