Crate utf8

Structs

BufReadDecoder

Wraps a std::io::BufRead buffered byte stream and decode it as UTF-8.

Incomplete

LossyDecoder

A push-based, lossy decoder for UTF-8. Errors are replaced with the U+FFFD replacement character.

Enums

BufReadDecoderError

DecodeError

Constants

REPLACEMENT_CHARACTER

The replacement character, U+FFFD. In lossy decoding, insert it for every decoding error.

Functions

decode