Struct utf8::BufReadDecoder

pub struct BufReadDecoder<B: BufRead> { /* private fields */ }

Wraps a std::io::BufRead buffered byte stream and decode it as UTF-8.

Implementations

impl<B: BufRead> BufReadDecoder<B>

pub fn read_to_string_lossy(buf_read: B) -> Result<String>

This is to Read::read_to_string what String::from_utf8_lossy is to String::from_utf8.

pub fn new(buf_read: B) -> Self

pub fn next_lossy(&mut self) -> Option<Result<&str>>

Same as BufReadDecoder::next_strict, but replace UTF-8 errors with U+FFFD.

pub fn next_strict(&mut self) -> Option<Result<&str, BufReadDecoderError<'_>>>

Decode and consume the next chunk of UTF-8 input.

This method is intended to be called repeatedly until it returns None, which represents EOF from the underlying byte stream. This is similar to Iterator::next, except that decoded chunks borrow the decoder (~iterator) so they need to be handled or copied before the next chunk can start decoding.